image_finder_backend/requirements.txt
selenium
requests
Pillow
flask
flask_cors
flask_restful
tqdm